requirements.txt
# Core dependencies
python-docx==1.1.0
Pillow==10.1.0
PyPDF2==3.0.1
pdf2image==1.16.3

# LLM API clients
openai==1.3.7
anthropic>=0.39.0
google-generativeai>=0.8.0
dashscope==1.14.1
google-api-core==2.17.0
grpcio==1.60.1
grpcio-status==1.60.1
protobuf==4.25.3
cryptography==41.0.7
cffi==1.15.1
psycopg2-binary==2.9.9

# Formula conversion
latex2mathml==3.77.0

# Configuration and utilities
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0

# Logging and monitoring
colorlog==6.8.0

# Image processing
opencv-python==4.8.1.78
numpy==1.26.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Optional: For better PDF handling
poppler-utils==0.1.0

# Web API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose==3.3.0
redis==5.0.4
httpx==0.27.0
lxml==4.9.3
